Move RouteAuth and protect login/register page
